+++ src/main/java/com/akiba/backend/auction/dto/response/BidHistoryListResponse.java
@@ -6,8 +6,11 @@
 package com.akiba.backend.auction.dto.response;
 
 import lombok.Builder;
+import java.math.BigDecimal;
 import lombok.Getter;
+import java.math.BigDecimal;
 import java.util.List;
+import java.math.BigDecimal;
 
 @Getter
 @Builder
@@ -15,10 +18,10 @@ public class BidHistoryListResponse {
 
     private Long postId;                // 경매 게시글 ID
     private int totalBidCount;          // 총 입찰 수
-    private int currentHighestBid;      // 현재 최고 입찰가
-    private int startPrice;             // 경매 시작가
-    private int bidStep;                // 입찰 단위
-    private int nextMinBid;             // 다음 최소 입찰 가능 금액
+    private BigDecimal currentHighestBid;      // 현재 최고 입찰가
+    private BigDecimal startPrice;             // 경매 시작가
+    private BigDecimal bidStep;                // 입찰 단위
+    private BigDecimal nextMinBid;             // 다음 최소 입찰 가능 금액
     private List<BidItem> bids;         // 입찰 목록 (가격 내림차순)
 
     @Getter
@@ -27,7 +30,7 @@ public class BidHistoryListResponse {
         private Long userId;
         private String nickname;
         private String profileImageUrl;
-        private int bidAmount;           // bidPrice → bidAmount
+        private BigDecimal bidAmount;           // bidPrice → bidAmount
         private String createdAt;
         private boolean isHighest;    // 현재 최고가 입찰인지
     }
